backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
faiss-cpu
sentence-transformers
requests
python-dotenv
numpy
